modules/storage/bucket/requirements.txt
aws-cdk-lib==2.79.1
constructs>=10.0.0,<11.0.0
cdk-nag==2.12.29
